requirement.txt
Flask
jsonrpclib-pelix
PyMySQL